requirements.txt
customtkinter==5.2.2
playwright==1.41.0
python-dateutil==2.8.2
